Refactor App and WebApp to enhance modularity by removing unused methods, adding state retrieval in AppInterface, and improving menu setup in AppMenu.
