+++ application/src/main/java/dev/ikm/komet/app/AppMenu.java
@@ -1,26 +1,43 @@
 package dev.ikm.komet.app;
 
+import de.jangassen.MenuToolkit;
+import de.jangassen.model.AppearanceMode;
 import dev.ikm.komet.app.aboutdialog.AboutDialog;
+import dev.ikm.komet.framework.graphics.Icon;
+import dev.ikm.komet.framework.preferences.KometPreferencesStage;
 import javafx.application.Platform;
+import javafx.event.ActionEvent;
+import javafx.scene.Scene;
 import javafx.scene.control.Menu;
 import javafx.scene.control.MenuBar;
 import javafx.scene.control.MenuItem;
+import javafx.scene.control.SeparatorMenuItem;
 import javafx.scene.input.KeyCode;
 import javafx.scene.input.KeyCodeCombination;
 import javafx.scene.input.KeyCombination;
 import javafx.scene.layout.BorderPane;
+import javafx.scene.layout.StackPane;
 import javafx.stage.Stage;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
 import java.io.IOException;
 import java.util.prefs.BackingStoreException;
 
+import static dev.ikm.komet.app.AppState.RUNNING;
 import static dev.ikm.komet.app.WebApp.IS_BROWSER;
+import static dev.ikm.komet.app.WebApp.IS_MAC_AND_NOT_TESTFX_TEST;
 import static dev.ikm.komet.kview.mvvm.view.changeset.exchange.GitTask.OperationMode.PULL;
 import static dev.ikm.komet.kview.mvvm.view.changeset.exchange.GitTask.OperationMode.SYNC;
 
 public class AppMenu {
 
+    private static final Logger LOG = LoggerFactory.getLogger(AppMenu.class);
+
     private final AppInterface app;
+    KometPreferencesStage kometPreferencesStage;
+
+    private static long windowCount = 1;
 
     public AppMenu(AppInterface app) {
         this.app = app;
@@ -148,4 +165,176 @@ public class AppMenu {
         classicKometMenuItem.setAccelerator(classicKometKeyCombo);
         return classicKometMenuItem;
     }
+
+
+    void setupMenus() {
+        Menu kometAppMenu;
+
+        if (IS_MAC_AND_NOT_TESTFX_TEST) {
+            MenuToolkit menuToolkit = MenuToolkit.toolkit();
+            kometAppMenu = menuToolkit.createDefaultApplicationMenu("Komet");
+        } else {
+            kometAppMenu = new Menu("Komet");
+        }
+
+        MenuItem prefsItem = new MenuItem("Komet preferences...");
+        prefsItem.setAccelerator(new KeyCodeCombination(KeyCode.COMMA, KeyCombination.META_DOWN));
+        prefsItem.setOnAction(event -> kometPreferencesStage.showPreferences());
+
+        if (IS_MAC_AND_NOT_TESTFX_TEST) {
+            kometAppMenu.getItems().add(2, prefsItem);
+            kometAppMenu.getItems().add(3, new SeparatorMenuItem());
+            MenuItem appleQuit = kometAppMenu.getItems().getLast();
+            appleQuit.setOnAction(event -> app.quit());
+        } else {
+            kometAppMenu.getItems().addAll(prefsItem, new SeparatorMenuItem());
+        }
+
+        MenuBar menuBar = new MenuBar(kometAppMenu);
+
+        if (app.getState().get() == RUNNING) {
+            Menu fileMenu = createFileMenu();
+            Menu editMenu = createEditMenu();
+            Menu viewMenu = createViewMenu();
+            menuBar.getMenus().addAll(fileMenu, editMenu, viewMenu);
+        }
+
+        if (IS_MAC_AND_NOT_TESTFX_TEST) {
+            MenuToolkit menuToolkit = MenuToolkit.toolkit();
+            menuToolkit.setApplicationMenu(kometAppMenu);
+            menuToolkit.setAppearanceMode(AppearanceMode.AUTO);
+            menuToolkit.setDockIconMenu(createDockMenu());
+            Menu windowMenu = createWindowMenuOnMacOS();
+            menuToolkit.autoAddWindowMenuItems(windowMenu);
+            menuToolkit.setGlobalMenuBar(menuBar);
+            menuToolkit.setTrayMenu(createSampleMenu());
+
+            // Add the window menu to the menu bar
+            menuBar.getMenus().add(windowMenu);
+        }
+
+        // Create and add the exchange menu to the menu bar
+        Menu exchangeMenu = createExchangeMenu();
+        menuBar.getMenus().add(exchangeMenu);
+
+        // Create and add the help menu to the menu bar
+        Menu helpMenu = createHelpMenu();
+        menuBar.getMenus().add(helpMenu);
+    }
+
+    private Menu createFileMenu() {
+        Menu fileMenu = new Menu("File");
+
+        // Import Dataset Menu Item
+        MenuItem importDatasetMenuItem = new MenuItem("Import Dataset...");
+        importDatasetMenuItem.setOnAction(actionEvent -> app.openImport(app.getPrimaryStage()));
+
+        // Export Dataset Menu Item
+        MenuItem exportDatasetMenuItem = new MenuItem("Export Dataset...");
+        exportDatasetMenuItem.setOnAction(actionEvent -> app.openExport(app.getPrimaryStage()));
+
+        // Add menu items to the File menu
+        fileMenu.getItems().addAll(importDatasetMenuItem, exportDatasetMenuItem);
+
+        if (IS_MAC_AND_NOT_TESTFX_TEST) {
+            // Close Window Menu Item
+            MenuToolkit tk = MenuToolkit.toolkit();
+            MenuItem closeWindowMenuItem = tk.createCloseWindowMenuItem();
+            fileMenu.getItems().addAll(new SeparatorMenuItem(), closeWindowMenuItem);
+        }
+
+        return fileMenu;
+    }
+
+    private Menu createEditMenu() {
+        Menu editMenu = new Menu("Edit");
+        editMenu.getItems().addAll(
+                createMenuItem("Undo"),
+                createMenuItem("Redo"),
+                new SeparatorMenuItem(),
+                createMenuItem("Cut"),
+                createMenuItem("Copy"),
+                createMenuItem("Paste"),
+                createMenuItem("Select All"));
+        return editMenu;
+    }
+
+    private Menu createViewMenu() {
+        Menu viewMenu = new Menu("View");
+        MenuItem classicKometMenuItem = new MenuItem("Classic Komet");
+        classicKometMenuItem.setAccelerator(new KeyCodeCombination(KeyCode.K, KeyCombination.SHORTCUT_DOWN));
+        classicKometMenuItem.setOnAction(actionEvent -> {
+            try {
+                app.getAppClassicKomet().launchClassicKomet();
+            } catch (IOException | BackingStoreException e) {
+                throw new RuntimeException(e);
+            }
+        });
+        viewMenu.getItems().add(classicKometMenuItem);
+        return viewMenu;
+    }
+
+    private Menu createWindowMenuOnMacOS() {
+        MenuToolkit menuToolkit = MenuToolkit.toolkit();
+        Menu windowMenu = new Menu("Window");
+        windowMenu.getItems().addAll(
+                menuToolkit.createMinimizeMenuItem(),
+                menuToolkit.createZoomMenuItem(),
+                menuToolkit.createCycleWindowsItem(),
+                new SeparatorMenuItem(),
+                menuToolkit.createBringAllToFrontItem());
+        return windowMenu;
+    }
+
+    private Menu createHelpMenu() {
+        Menu helpMenu = new Menu("Help");
+        helpMenu.getItems().add(new MenuItem("Getting started"));
+        return helpMenu;
+    }
+
+
+    private MenuItem createMenuItem(String title) {
+        MenuItem menuItem = new MenuItem(title);
+        menuItem.setOnAction(this::handleEvent);
+        return menuItem;
+    }
+
+    private Menu createDockMenu() {
+        Menu dockMenu = createSampleMenu();
+        MenuItem open = new MenuItem("New Window");
+        open.setGraphic(Icon.OPEN.makeIcon());
+        open.setOnAction(e -> createNewStage());
+        dockMenu.getItems().addAll(new SeparatorMenuItem(), open);
+        return dockMenu;
+    }
+
+    private Menu createSampleMenu() {
+        Menu trayMenu = new Menu();
+        trayMenu.setGraphic(Icon.TEMPORARY_FIX.makeIcon());
+        MenuItem reload = new MenuItem("Reload");
+        reload.setGraphic(Icon.SYNCHRONIZE_WITH_STREAM.makeIcon());
+        reload.setOnAction(this::handleEvent);
+        MenuItem print = new MenuItem("Print");
+        print.setOnAction(this::handleEvent);
+
+        Menu share = new Menu("Share");
+        MenuItem mail = new MenuItem("Mail");
+        mail.setOnAction(this::handleEvent);
+        share.getItems().add(mail);
+
+        trayMenu.getItems().addAll(reload, print, new SeparatorMenuItem(), share);
+        return trayMenu;
+    }
+
+    private void handleEvent(ActionEvent actionEvent) {
+        LOG.debug("clicked " + actionEvent.getSource());  // NOSONAR
+    }
+
+    private static void createNewStage() {
+        Stage stage = new Stage();
+        stage.setScene(new Scene(new StackPane()));
+        stage.setTitle("New stage" + " " + (windowCount++));
+        stage.show();
+    }
+
 }
